+++ example/lib/examples/examples.dart
@@ -28,3 +28,5 @@ export 'todo.dart';
 export 'transform.dart';
 export 'test_graphic_resizing.dart';
 export 'test_memory_cleanup.dart';
+export 'test_webgl_context_lifecycle.dart';
+export 'test_webgl_scroll_recycling.dart';